Feat: upgrade to 1.4.1
