+++ src/components/tx-flow/flows/UpdateSafe/UpdateSafeReview.tsx
@@ -5,7 +5,7 @@ import ExternalLink from '@/components/common/ExternalLink'
 import { useCurrentChain } from '@/hooks/useChains'
 import useSafeInfo from '@/hooks/useSafeInfo'
 import { createUpdateSafeTxs } from '@/services/tx/safeUpdateParams'
-import { createMultiSendCallOnlyTx } from '@/services/tx/tx-sender'
+import { createMultiSendCallOnlyTx, createTx } from '@/services/tx/tx-sender'
 import { SafeTxContext } from '../../SafeTxProvider'
 import SignOrExecuteForm from '@/components/tx/SignOrExecuteForm'
 import useAsync from '@/hooks/useAsync'
@@ -24,7 +24,8 @@ export const UpdateSafeReview = () => {
     }
 
     const txs = await createUpdateSafeTxs(safe, chain)
-    createMultiSendCallOnlyTx(txs).then(setSafeTx).catch(setSafeTxError)
+    const safeTxPromise = txs.length > 1 ? createMultiSendCallOnlyTx(txs) : createTx(txs[0])
+    safeTxPromise.then(setSafeTx).catch(setSafeTxError)
   }, [safe, safeLoaded, chain, setSafeTx, setSafeTxError])
 
   return (
