+++ src/utils/transactions.ts
@@ -21,24 +21,18 @@ import {
 } from './transaction-guards'
 import type { MetaTransactionData } from '@safe-global/safe-core-sdk-types/dist/src/types'
 import { OperationType } from '@safe-global/safe-core-sdk-types/dist/src/types'
-import { getReadOnlyGnosisSafeContract, isValidMasterCopy } from '@/services/contracts/safeContracts'
+import { getReadOnlyGnosisSafeContract } from '@/services/contracts/safeContracts'
 import extractTxInfo from '@/services/tx/extractTxInfo'
 import type { AdvancedParameters } from '@/components/tx/AdvancedParams'
 import type { SafeTransaction, TransactionOptions } from '@safe-global/safe-core-sdk-types'
 import { FEATURES, hasFeature } from '@/utils/chains'
 import uniqBy from 'lodash/uniqBy'
 import { Errors, logError } from '@/services/exceptions'
-import { Safe_to_l2_migration__factory } from '@/types/contracts'
 import { type BaseTransaction } from '@safe-global/safe-apps-sdk'
 import { isEmptyHexData } from '@/utils/hex'
-import { type ExtendedSafeInfo } from '@/store/safeInfoSlice'
-import { getSafeContractDeployment } from '@/services/contracts/deployments'
-import { sameAddress } from './addresses'
 import { isMultiSendCalldata } from './transaction-calldata'
 import { decodeMultiSendData } from '@safe-global/protocol-kit/dist/src/utils'
-import { __unsafe_createMultiSendTx } from '@/services/tx/tx-sender'
 import { getOriginPath } from './url'
-import { getSafeToL2MigrationDeployment } from '@safe-global/safe-deployments'
 
 export const makeTxFromDetails = (txDetails: TransactionDetails): Transaction => {
   const getMissingSigners = ({
@@ -235,120 +229,6 @@ export const isTrustedTx = (tx: TransactionSummary) => {
 
 export const isImitation = ({ txInfo }: TransactionSummary): boolean => {
   return isTransferTxInfo(txInfo) && isERC20Transfer(txInfo.transferInfo) && Boolean(txInfo.transferInfo.imitation)
-}
-
-/**
- *
- * If the Safe is using a invalid masterCopy this function will modify the passed in `safeTx` by making it a MultiSend that migrates the Safe to L2 as the first action.
- *
- * This only happens under the conditions that
- * - The Safe's nonce is 0
- * - The SafeTx's nonce is 0
- * - The Safe is using an invalid masterCopy
- * - The SafeTx is not already including a Migration
- *
- * @param safeTx original SafeTx
- * @param safe
- * @param chain
- * @returns
- */
-export const prependSafeToL2Migration = (
-  safeTx: SafeTransaction | undefined,
-  safe: ExtendedSafeInfo,
-  chain: ChainInfo | undefined,
-): Promise<SafeTransaction | undefined> => {
-  if (!chain) {
-    throw new Error('No Network information available')
-  }
-
-  const safeL2Deployment = getSafeContractDeployment(chain, safe.version)
-  const safeL2DeploymentAddress = safeL2Deployment?.networkAddresses[chain.chainId]
-  const safeToL2MigrationDeployment = getSafeToL2MigrationDeployment({ network: chain.chainId })
-  const safeToL2MigrationAddress = safeToL2MigrationDeployment?.networkAddresses[chain.chainId]
-
-  if (
-    !safeTx ||
-    safeTx.signatures.size > 0 ||
-    !chain.l2 ||
-    safeTx.data.nonce > 0 ||
-    isValidMasterCopy(safe.implementationVersionState) ||
-    !safeToL2MigrationAddress ||
-    !safeL2DeploymentAddress
-  ) {
-    // We do not migrate on L1s
-    // We cannot migrate if the nonce is > 0
-    // We do not modify already signed txs
-    // We do not modify supported masterCopies
-    // We cannot migrate if no migration contract or L2 contract exists
-    return Promise.resolve(safeTx)
-  }
-
-  const safeToL2MigrationInterface = Safe_to_l2_migration__factory.createInterface()
-
-  if (sameAddress(safe.implementation.value, safeL2DeploymentAddress)) {
-    // Safe already has the correct L2 masterCopy
-    // This should in theory never happen if the implementationState is valid
-    return Promise.resolve(safeTx)
-  }
-
-  // If the Safe is a L1 masterCopy on a L2 network and still has nonce 0, we prepend a call to the migration contract to the safeTx.
-  const txData = safeTx.data.data
-
-  let internalTxs: MetaTransactionData[]
-  if (isMultiSendCalldata(txData)) {
-    // Check if the first tx is already a call to the migration contract
-    internalTxs = decodeMultiSendData(txData)
-  } else {
-    internalTxs = [{ to: safeTx.data.to, operation: safeTx.data.operation, value: safeTx.data.value, data: txData }]
-  }
-
-  if (sameAddress(internalTxs[0]?.to, safeToL2MigrationAddress)) {
-    // We already migrate. Nothing to do.
-    return Promise.resolve(safeTx)
-  }
-
-  // Prepend the migration tx
-  const newTxs: MetaTransactionData[] = [
-    {
-      operation: 1, // DELEGATE CALL REQUIRED
-      data: safeToL2MigrationInterface.encodeFunctionData('migrateToL2', [safeL2DeploymentAddress]),
-      to: safeToL2MigrationAddress,
-      value: '0',
-    },
-    ...internalTxs,
-  ]
-
-  return __unsafe_createMultiSendTx(newTxs)
-}
-
-export const extractMigrationL2MasterCopyAddress = (safeTx: SafeTransaction | undefined): string | undefined => {
-  if (!safeTx) {
-    return undefined
-  }
-
-  if (!isMultiSendCalldata(safeTx.data.data)) {
-    return undefined
-  }
-
-  const innerTxs = decodeMultiSendData(safeTx.data.data)
-  const firstInnerTx = innerTxs[0]
-  if (!firstInnerTx) {
-    return undefined
-  }
-
-  const safeToL2MigrationDeployment = getSafeToL2MigrationDeployment()
-  const safeToL2MigrationAddress = safeToL2MigrationDeployment?.defaultAddress
-  const safeToL2MigrationInterface = Safe_to_l2_migration__factory.createInterface()
-
-  if (
-    firstInnerTx.data.startsWith(safeToL2MigrationInterface.getFunction('migrateToL2').selector) &&
-    sameAddress(firstInnerTx.to, safeToL2MigrationAddress)
-  ) {
-    const callParams = safeToL2MigrationInterface.decodeFunctionData('migrateToL2', firstInnerTx.data)
-    return callParams[0]
-  }
-
-  return undefined
 }
 
 export const getSafeTransaction = async (safeTxHash: string, chainId: string, safeAddress: string) => {
